Fixed crash in XamlBind sample by changing binding
Changed a binding in OtherBindings.xaml to fix a crash. Previously the FontSize of a TextBlock was bound to Model.IntPropWithINPC, which caused a crash when that value went to zero (by dragging the corresponding slider all the way to the left).

With this change the FontSize is bound instead to the text length, and doesn't crash even when the length goes to zero -- not sure why that is the case, tested in Desktop build 10240.